Improve JUnit test cases
