+++ src/test/java/evelyn/task/TaskTest.java
@@ -3,6 +3,7 @@ package evelyn.task;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class TaskTest {
 
     @Test
@@ -12,9 +13,9 @@ public class TaskTest {
     }
 
     @Test
-    public void checkedTodoTest() {
-        Todo todo = new Todo("Test 2", true);
-        assertEquals("[T][X] Test 2", todo.toString());
+    public void longTodoTest() {
+        Todo todo = new Todo("This is a long todo task message", false);
+        assertEquals("[T][ ] This is a long todo task message", todo.toString());
     }
 
     @Test
@@ -23,11 +24,6 @@ public class TaskTest {
         assertEquals("[D][ ] Test 3 (by: 2020-03-11)", deadline.toString());
     }
 
-    @Test
-    public void checkedDeadlineTest() {
-        Deadline deadline = new Deadline("Test 4", "2020-03-11", true);
-        assertEquals("[D][X] Test 4 (by: 2020-03-11)", deadline.toString());
-    }
 
     @Test
     public void uncheckedEventTest() {
@@ -35,9 +31,4 @@ public class TaskTest {
         assertEquals("[E][ ] Test 5 (from: 2020-04-04 to: 2020-04-04)", event.toString());
     }
 
-    @Test
-    public void checkedEventTest() {
-        Event event = new Event("Test 6", "2020-04-04", "2020-04-04", true);
-        assertEquals("[E][X] Test 6 (from: 2020-04-04 to: 2020-04-04)", event.toString());
-    }
 }
